3.4-django-testing/django_testing/requirements.txt
pip install pytest pytest-django model_bakery factory-boydjango
djangorestframework
django-filter
psycopg2-binary
model_bakery